+++ node/input.d.ts
@@ -1,5 +1,5 @@
-import { Stream } from 'stream';
-import { Blob } from 'buffer';
+import type { Stream } from 'stream';
+import type { Blob } from 'buffer';
 
 // FYI: `Buffer` is also an `ArrayBuffer`.
 export type Input = string | Stream | Blob | Buffer;