+++ source/export/readXlsxFileWebWorker.js
@@ -7,7 +7,7 @@ import parseSpreadsheetContents from '../xlsx/parseSpreadsheetContents.js'
  * Reads an `.xlsx` file.
  * @param  {(File|Blob|ArrayBuffer)} input
  * @param  {object} [options]
- * @return {Promise<ReadFileResult>}
+ * @return {Promise<Sheet[]>}
  */
 export default function readXlsxFile(file, options) {
 	return unpackXlsxFile(file)