+++ types/Options.d.ts
@@ -1,0 +1,5 @@
+export interface Options<ParsedNumber = number> {
+	trim?: boolean;
+	parseNumber?: (string: string) => ParsedNumber;
+	dateFormat?: string;
+}