+++ types/Sheet.d.ts
@@ -1,0 +1,6 @@
+import type { SheetData } from './SheetData.d.js'
+
+export type Sheet<ParsedNumber = number> = {
+	sheet: string;
+	data: SheetData<ParsedNumber>;
+}